refactor: extract project-file-watch.js from project.js (PR-05)
Move file and directory watcher functions (startFileWatch, stopFileWatch,
startDirWatch, stopDirWatch, stopAllDirWatches) into a dedicated module.
project.js reduced from 4,197 to 4,120 lines.
